+++ src/schemas/ba-sak/begrunnelse/nasjonaleTriggere/utils.ts
@@ -1,14 +1,7 @@
-import { BegrunnelseDokumentNavn } from '../../../../util/typer';
-import { Behandlingstema } from '../typer';
-
-export const erNasjonalBegrunnelse = document =>
-  document[BegrunnelseDokumentNavn.BEHANDLINGSTEMA] &&
-  document[BegrunnelseDokumentNavn.BEHANDLINGSTEMA].includes(Behandlingstema.NASJONAL);
+import { hentNasjonaltFeltRegler } from '../utils';
 
 export const hentNasjonaleTriggereRegler = rule =>
-  rule.custom((currentValue, { document }) => {
-    if (erNasjonalBegrunnelse(document) && currentValue !== undefined) {
-      return 'En nasjonal begrunnelse-trigger er valgt, men behandlingstema for begrunnelsen er ikke nasjonal.';
-    }
-    return true;
-  });
+  hentNasjonaltFeltRegler(
+    rule,
+    'En nasjonal begrunnelse-trigger er valgt, men behandlingstema for begrunnelsen er ikke nasjonal.',
+  );
